warning fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@11093 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/VKontakte/src/vk_files.cpp
@@ -230,7 +230,7 @@ void CVkProto::OnReciveUploadServer(NETLIBHTTPREQUEST *reply, AsyncHttpRequest *
 	memcpy(pData, (void *)DataEnd.GetBuffer(), DataEnd.GetLength());
 	// } Body
 
-	pUploadReq->	dataLength = dataLength;
+	pUploadReq->dataLength = (int)dataLength;
 	pUploadReq->pUserInfo = pReq->pUserInfo;
 	Push(pUploadReq);
 }
